+++ docs/CHANGE_LOG.md
@@ -5,12 +5,19 @@
 #### **New Features:**
 - Added full binding support for "data-syntax-tab-contents", which allows some configuration to be set per tab.
 
+#### **Binding Options Tab Contents:**
+- Added a new binding option called "title", which states the title to use for a Tab (defaults to null, which forces the language name ot be used).
+
+#### **Binding Options Tab Contents - Custom Triggers:**
+- Added a new binding custom trigger called "onOpen", which states an event that should be triggered when a tab is opened (passes to the language to the event).
+
 ### **CSS:**
 - Renamed the CSS class "label" to "language-label" (much clearer).
 
 #### **Documentation:**
 - Minor documentation improvements.
-- Added new binding documentation file "BINDING_TAB_CONTENT_OPTIONS.md".
+- Added new binding documentation file "binding/tabs/TAB_CONTENTS.md".
+- Added new binding documentation file "binding/tabs/CUSTOM_TRIGGERS.md".
 - Some of the documentation has been renamed and moved into category folders (the filenames were getting confusing).
 
 <br>